refactor: refactor OnboardingForm, deprecate Formik and reorganize the folder
