+++ src/components/forms/index.ts
@@ -1,5 +1,4 @@
 export * from "./connector";
 export * from "./model";
 export * from "./pipeline";
-export * from "./mgmt";
 export * from "./commons";
