+++ src/components/onboarding/OnboardingForm/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./OnboardingForm";
+export * from "./OnboardingForm";
